rewrite getting person fractie to getting mandataris fracties
